refactor: let the prompts layer choose the generation failure message
The action was reading `getError('pluginConfig')` and `getError('sdkRepos')` to
decide which message to show — knowing the shape of an error payload in order to
pick wording. Hand the whole `ServiceError` to one prompt, as `sdk generate`
does, and let the prompts class hold the variations. Portal branches in its
action only because its error is a union and it needs `instanceof` to tell a
`ServiceError` from a stream; ours is not.

Drop the guard around the save so it reads like portal's two lines, and with it
the `pluginSaveFailed` prompt that then had no caller. Note this restores the
raw stack trace a payload that is not a zip produces.
